+++ Shared/Models/DataSourceType.swift
@@ -8,10 +8,6 @@
 import Foundation
 import GRDB
 
-enum DataSourceDisplayMode: String, Codable, Hashable {
-    case count, sum
-}
-
 enum DataSourceType {
     case empty
     case entries(datasetID: DataSet.ID, mode: DataSourceDisplayMode)
@@ -317,17 +313,5 @@ extension DataSourceType: DatabaseValueConvertible {
         }
         
         return try? Database.jsonConverter.decode(DataSourceType.self, from: data)
-    }
-}
-
-extension DataSourceDisplayMode: DatabaseValueConvertible {
-    var databaseValue: DatabaseValue { "".databaseValue }
-    
-    static func fromDatabaseValue(_ dbValue: DatabaseValue) -> DataSourceDisplayMode? {
-        guard let str = String.fromDatabaseValue(dbValue), let data = str.data(using: .utf8) else {
-            return nil
-        }
-        
-        return try? Database.jsonConverter.decode(DataSourceDisplayMode.self, from: data)
     }
 }
